Make exception mapper filters public
Motivation:

Some users want to map exceptions earlier in the request processing
pipeline in order to observe a real HTTP status code in other filters.

Modifications:

- Move `DefaultHttpServerBuilder.ExceptionMapperServiceFilter` ->
`io.servicetalk.http.api.HttpExceptionMapperServiceFilter`;
- Move `io.servicetalk.grpc.netty.CatchAllHttpServiceFilter` ->
`io.servicetalk.grpc.api.GrpcExceptionMapperServiceFilter`;
- Rename `CatchAllHttpServiceFilterTest` ->
`GrpcExceptionMapperServiceFilterTest`;
- Remove `io.servicetalk.grpc.netty.GrpcUtils` duplicate;

Result:

Users have access to our exception mapper implementations and can move
them around other filters in the processing pipeline.
