+++ servicetalk-grpc-api/src/main/java/io/servicetalk/grpc/api/GrpcExceptionMapperServiceFilter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2021 Apple Inc. and the ServiceTalk project authors
+ * Copyright © 2021-2022 Apple Inc. and the ServiceTalk project authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.servicetalk.grpc.netty;
+package io.servicetalk.grpc.api;
 
 import io.servicetalk.concurrent.api.Single;
 import io.servicetalk.http.api.HttpExecutionStrategies;
@@ -28,13 +28,22 @@ import io.servicetalk.http.api.StreamingHttpServiceFilterFactory;
 
 import static io.servicetalk.concurrent.api.Single.succeeded;
 import static io.servicetalk.grpc.api.GrpcHeaderValues.APPLICATION_GRPC;
-import static io.servicetalk.grpc.netty.GrpcUtils.newErrorResponse;
+import static io.servicetalk.grpc.api.GrpcUtils.newErrorResponse;
 
-final class CatchAllHttpServiceFilter implements StreamingHttpServiceFilterFactory {
+/**
+ * Filter that maps known {@link Throwable} subtypes into an gRPC response with an appropriate {@link GrpcStatusCode}.
+ * <p>
+ * This filter is recommended to be placed as early as possible to make sure it captures all exceptions that may be
+ * generated by other filters.
+ */
+public final class GrpcExceptionMapperServiceFilter implements StreamingHttpServiceFilterFactory {
 
-    public static final StreamingHttpServiceFilterFactory INSTANCE = new CatchAllHttpServiceFilter();
+    /**
+     * Instance of {@link GrpcExceptionMapperServiceFilter}.
+     */
+    public static final StreamingHttpServiceFilterFactory INSTANCE = new GrpcExceptionMapperServiceFilter();
 
-    private CatchAllHttpServiceFilter() {
+    private GrpcExceptionMapperServiceFilter() {
         // Singleton
     }
