+++ servicetalk-grpc-netty/src/test/java/io/servicetalk/grpc/netty/GrpcExceptionMapperServiceFilterTest.java
@@ -15,14 +15,16 @@
  */
 package io.servicetalk.grpc.netty;
 
+import io.servicetalk.grpc.api.GrpcExceptionMapperServiceFilter;
+
 import org.junit.jupiter.api.Test;
 
 import static io.servicetalk.http.netty.AsyncContextHttpFilterVerifier.verifyServerFilterAsyncContextVisibility;
 
-class CatchAllHttpServiceFilterTest {
+class GrpcExceptionMapperServiceFilterTest {
 
     @Test
     void verifyAsyncContext() throws Exception {
-        verifyServerFilterAsyncContextVisibility(CatchAllHttpServiceFilter.INSTANCE);
+        verifyServerFilterAsyncContextVisibility(GrpcExceptionMapperServiceFilter.INSTANCE);
     }
 }
